refactor(shared-contracts): move outcome archetype type to stable shared surface
- add OutcomeArchetypeId to shared simulation runtime contracts
- switch simulation, reporting, persistence, UI, and story imports to use shared contracts
- remove type coupling from UI/reporting to simulation rule modules
- keep runtime behavior unchanged (type-surface refactor only)
