+++ docs/CODE_DOCUMENTATION_PASS_TRACKER.md
@@ -250,15 +250,15 @@ Start date: 2026-04-20
 | F-002 | simulation | Orchestration concentration in one service may reduce readability/testability | src/domains/simulation/services/simulation_runner.ts | 5 | 3 | 1.67 | medium | schedule | unassigned | open |
 | F-003 | ui | Mixed component filename casing may reduce consistency | src/ui/components | 2 | 2 | 1.00 | low | watch | unassigned | open |
 | F-004 | simulation | Repeated score/stakeholder change conversion helpers across rule files suggest shared utility opportunity | src/domains/simulation/rules/resolve_action.ts | 3 | 2 | 1.50 | low | watch | unassigned | open |
-| F-005 | ui | UI imports outcome archetype types from simulation rule modules, suggesting contract leakage that could be moved to a shared reporting/runtime type surface | src/ui/components/results/share_result_card.vue | 3 | 2 | 1.50 | low | watch | unassigned | open |
-| F-006 | ui | Main game store centralizes many concerns (content loading, run lifecycle, persistence, tutorial triggers, modal state), increasing maintenance cost | src/ui/stores/game_store.ts | 4 | 3 | 1.33 | medium | schedule | unassigned | in_progress |
+| F-005 | ui | UI imports outcome archetype types from simulation rule modules, suggesting contract leakage that could be moved to a shared reporting/runtime type surface | src/ui/components/results/share_result_card.vue | 3 | 2 | 1.50 | low | watch | unassigned | completed |
+| F-006 | ui | Main game store centralizes many concerns (content loading, run lifecycle, persistence, tutorial triggers, modal state), increasing maintenance cost | src/ui/stores/game_store.ts | 4 | 3 | 1.33 | medium | schedule | unassigned | completed |
 
 ## Boundary Integrity Checks
 
 | Check | Area | Result | Notes |
 |---|---|---|---|
 | simulation imports UI/browser APIs | simulation | pass | code usage check passed; only comment mention in engine.ts |
-| UI contains gameplay rule logic | ui | pass | final sweep confirms no direct rule execution; only limited type-only imports from simulation rule modules |
+| UI contains gameplay rule logic | ui | pass | final sweep confirms no direct rule execution and no direct UI imports from simulation rule modules |
 | persistence modifies gameplay behavior | persistence | pass | persistence modules are serialization/validation/adapters only |
 | reporting modifies simulation state | reporting | pass | reporting modules are payload building/encoding only |
 
@@ -270,5 +270,5 @@ Start date: 2026-04-20
 
 ## Next Actions
 
-1. Execute WS-3 / F-005 from docs/CLEANUP_REFACTOR_IMPLEMENTATION_PLAN.md (move archetype types to stable shared contract surface)
+1. Execute WS-4 / F-004 from docs/CLEANUP_REFACTOR_IMPLEMENTATION_PLAN.md (consolidate duplicated rule conversion helpers)
 2. Assign owners for WS-1, WS-2, and verification gates
