+++ src/ui/views/end_of_run_view.vue
@@ -208,7 +208,7 @@
 import { computed, onMounted, ref } from 'vue'
 import { useRouter } from 'vue-router'
 import { useGameStore } from '@/ui/stores/game_store'
-import type { OutcomeArchetypeId } from '@/domains/simulation/rules'
+import type { OutcomeArchetypeId } from '@/shared/contracts'
 import AboutModal from '@/ui/components/common/about_modal.vue'
 import RulesModal from '@/ui/components/common/rules_modal.vue'
 import ShareResultCard from '@/ui/components/results/share_result_card.vue'
